+++ CHANGES.rst
@@ -30,6 +30,7 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 Added
 -----
 
+* Added CHOOSE() function
 * Added ISBLANK() function
 * Added ISLOGICAL() function
 * Added ISNONTEXT() function